@@ -17,10 +17,10 @@
 //= require dataTables/jquery.dataTables
 //= require dataTables/bootstrap/3/jquery.dataTables.bootstrap
 //= require bootstrap-datepicker
-//= require csv_preview
 //
 // Required by Blacklight
 //= require blacklight/blacklight
+//= require csv_preview
 
 //= require_tree .
 //= require hyrax
