@@ -1,6 +1,9 @@
-$( document ).on('turbolinks:load', function() {
+Blacklight.onLoad(function() {
     $('.csv-preview').each(function() {
         var preview = $(this);
+        if(preview.hasClass('done')) {
+            return true
+        }
         $.ajax({
             url: preview.data('url'),
             success: function (data) {
@@ -22,6 +25,7 @@ $( document ).on('turbolinks:load', function() {
                 } else {
                     preview.find('.csv-preview-size').hide();
                 }
+                preview.addClass('done')
             },
             error: function() {
                 preview.find('.csv-preview-error').show();
