@@ -1,6 +1,9 @@
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
+  # Added for Rails 5.2
+  config.secret_key_base = ENV['SECRET_KEY_BASE_PRODUCTION']
+
   # Code is not reloaded between requests.
   config.cache_classes = true
 
@@ -59,7 +62,9 @@ Rails.application.configure do
     config.force_ssl = false
   else
     config.force_ssl = true #default if nothing specified is more secure.
-    Rails.application.routes.default_url_options = {protocol: 'https', host: ENV['MDR_HOST']}
+    Rails.application.routes.default_url_options = \
+      Hyrax::Engine.routes.default_url_options = \
+      {protocol: 'https', host: ENV['MDR_HOST']}
   end
 
   # Use the lowest log level (:debug) to ensure availability of diagnostic information
@@ -97,7 +102,7 @@ Rails.application.configure do
 
   # Enable locale fallbacks for I18n (makes lookups for any locale fall back to
   # the I18n.default_locale when a translation cannot be found).
-  config.i18n.fallbacks = true
+  # config.i18n.fallbacks = true
 
   # Send deprecation notices to registered listeners.
   config.active_support.deprecation = :notify
@@ -114,7 +119,7 @@ Rails.application.configure do
   }
 
   config.middleware.use ExceptionNotification::Rack,
-    ignore_if: lambda { |env, exception| !env.nil? },
+    ignore_exceptions: ['I18n::InvalidLocale', 'Riiif::ConversionError'] + ExceptionNotifier.ignored_exceptions,
     error_grouping: true,
     email: {
       email_prefix: "[MDR #{ENV['ERROR_NOTIFICATION_SUBJECT_PREFIX']}] ",
