@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20200108111832) do
+ActiveRecord::Schema.define(version: 2020_01_08_111832) do
 
   create_table "bookmarks", force: :cascade do |t|
     t.integer "user_id", null: false
