@@ -32,6 +32,49 @@ RSpec.describe ExportsController do
           expect(status).to eql(200)
           expect(json['columns']).to match_array(["Code", "Study_participation", "Census_usually_resident_population_count"])
           expect(json['data']).to match_array([
+            ["1", "Full-time study", "1000911"],
+            ["2", "Part-time study", "149919"],
+            ["4", "Not studying", "3548922"],
+            ["7", "Response not identifiable", "0"],
+            ["9", "Not stated", "0"],
+            ["TotalStated", "Total stated", "4699755"],
+            ["Total", "Total", "4699755"]
+          ])
+          expect(json['total_rows']).to eql(7)
+          expect(json['maximum_rows']).to eql(100)
+          expect(json['file_name']).to eql("example.csv")
+        end
+      end
+
+      describe 'authentication' do
+        let(:file_set) { create(:file_set, :authenticated, content: File.open(fixture_path + '/csv/example.csv')) }
+
+        context 'unauthenticated' do
+          it 'should return an unauthenticated error' do
+            expect(status).to eql(401)
+          end
+        end
+
+        context 'authenticated' do
+          let(:user) { create(:user) }
+          before do
+            sign_in user
+          end
+
+          it 'should return success' do
+            expect(status).to eql(200)
+          end
+        end
+      end
+    end
+
+    context 'tsv' do
+      context 'open' do
+        let(:file_set) { create(:file_set, :open, content: File.open(fixture_path + '/tsv/example.tsv')) }
+        it 'should return a json export of the TSV file' do
+          expect(status).to eql(200)
+          expect(json['columns']).to match_array(["Code", "Study_participation", "Census_usually_resident_population_count"])
+          expect(json['data']).to match_array([
                                                   ["1", "Full-time study", "1000911"],
                                                   ["2", "Part-time study", "149919"],
                                                   ["4", "Not studying", "3548922"],
@@ -42,12 +85,12 @@ RSpec.describe ExportsController do
                                               ])
           expect(json['total_rows']).to eql(7)
           expect(json['maximum_rows']).to eql(100)
-          expect(json['file_name']).to eql("example.csv")
+          expect(json['file_name']).to eql("example.tsv")
         end
       end
 
       describe 'authentication' do
-        let(:file_set) { create(:file_set, :authenticated, content: File.open(fixture_path + '/csv/example.csv')) }
+        let(:file_set) { create(:file_set, :authenticated, content: File.open(fixture_path + '/tsv/example.tsv')) }
 
         context 'unauthenticated' do
           it 'should return an unauthenticated error' do
