@@ -6,9 +6,9 @@ RSpec.describe UserAuthorisationService do
 
   context 'without env vars' do
     before do
-      allow(ENV).to receive(:[]).with('USER_AUTHORISATION_LDAP_HOST').and_return(nil)
-      allow(ENV).to receive(:[]).with('USER_AUTHORISATION_LDAP_BASE').and_return(nil)
-      allow(ENV).to receive(:[]).with('USER_AUTHORISATION_LDAP_ATTRIBUTE').and_return(nil)
+      allow(ENV).to receive(:[]).with('LDAP_HOST').and_return(nil)
+      allow(ENV).to receive(:[]).with('LDAP_BASE').and_return(nil)
+      allow(ENV).to receive(:[]).with('LDAP_ATTRIBUTE').and_return(nil)
     end
     it { expect(service.enabled?).to be false }
     it { expect(service.update_attributes).to be false }
@@ -16,9 +16,9 @@ RSpec.describe UserAuthorisationService do
 
   context 'with env vars' do
     before do
-      allow(ENV).to receive(:[]).with('USER_AUTHORISATION_LDAP_HOST').and_return('localhost')
-      allow(ENV).to receive(:[]).with('USER_AUTHORISATION_LDAP_BASE').and_return('ou=people,dc=test,dc=com')
-      allow(ENV).to receive(:[]).with('USER_AUTHORISATION_LDAP_ATTRIBUTE').and_return('uid')
+      allow(ENV).to receive(:[]).with('LDAP_HOST').and_return('localhost')
+      allow(ENV).to receive(:[]).with('LDAP_BASE').and_return('ou=people,dc=test,dc=com')
+      allow(ENV).to receive(:[]).with('LDAP_ATTRIBUTE').and_return('uid')
     end
     it { expect(service.enabled?).to be true }
 
